requirements.txt
wand
ffmpeg-python
mediainfo
